requirements.txt
PyYAML
Pillow
wheel
build
twine
pylint
yapf
tqdm
tensorflow
tensorflow-addons
pandas
pytest
pytest-cov
setuptools